Update Dashboard Admin dan halaman mahasiswa

+++ database/migrations/2025_04_22_094214_create_mahasiswa_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('nama');
-            $table->integer('nim')->unique();
+            $table->bigInteger('nim')->unique();
             $table->string('jurusan');
             $table->string('semester');
             $table->text('bio')->nullable();
